More around portal + new "partial" request handling - Portal controller with fully working page display elements. Ready for editor. - Add few simple item types like nav, userinfo (logged in user links), contact links - Add KB type with both content and sidebar section handlers
System request class has new method of identifying 'partial' requests. To request a partial request, send '_partial' in the query or post data. If the action is capable of handling partial requests, then the page will be served without the full layout (can be stuck in the same page etc).