Extends ticket search for reports system
We need to be able to use the functionlity of the ticket search but with the
ability to customise the generated queries. This provides this by
asking the ticket search for a QUeryBuilder instance that is setup with the
correct base query. The report queries are now able to simply add
to this base query