Edit Doctrine to support custom ProxyFactory needed for better implementation of our static ReflectionProperty hack.
Doctrine devs say they will not open the API to support customizations to the proxy factory, so it's a bit of a workaround to implement the customizations needed for us. Due to the private nature of these classes, the only way is to copy their classes exactly and make the two minor changes needed.