Better detection of URL rewriting.
New system script to test for URL rewriting with a special URL:
    /__checkurlrewrite/path

Three possible outcomes:
1) The file fails to load (404). This means URL rewriting is
not available.
2) The file loads. This means URL rewriting is available.
2a) The PHP file compares the base path with what it should be
to detect if RewriteBase is correct.

URL rewriting is not turned on unless (2a) is reached and correct.

A new section to alert the user of possible problems exists on the
checks page.