All entities now use notify change tracking.
Went through each entity and made changes where required to support
the notify change tracking policy.

To enable the policy, the entity must have a metadata definition with:

	$metadata->setChangeTrackingPolicy(ClassMetadataInfo::CHANGETRACKING_NOTIFY);

Then whenever a model field is changed from within the class, you must
register the change:

	$this->_onPropertyChanged('person', $old_value, $value);

Or use the shortcuts:

	$this->setModelField('person', $value);
	$this['person'] = $value;

If you explicitly reference a model field, you are bypassing the notify
system and the object will NOT be scheduled for updates!

	// This is wrong, it wont persist
	$this->person = $value;