Bypass the Doctrine EM when saving status of email source when an error might've happened.
Doctrine cannot recover from errors during commit-time in the UoW (e.g., fatal connection errors,
invalid FK's, etc). Since Doctrine can't know how to reset the state of the UoW when there's an error
here, it closes the em so we can't use it to update the source status to say there was an error. Simple
fix is to issue a "plain old" db update which will work even if the em is closed.