- Add order direction (ASC/DESC) support  - Add preliminary support for converting DPQL to SQL. This does not do any processing on functions/placeholders/tables yet.
Work to analyze parts of the query and determine elements that need to be pushed up to the select level (like fields in a group/split) needs to be done. Functions must also be analyzed as they may not print out SQL (eg, X() and Y()). Placeholders need to be analzed in regards to the stack (for the parent comparison) as it will manipulate the outcome. Field references need to be analyzed to determine validity and to resolve joins.

The technical design for this is currently unknown and is a really big challenge to determine.