Add original_subject field to aid in subject matching.
This is for certain cases where a user might reply to an existing ticket but the originally submitted ticket has a subject that it wouldn't match.

For example, UserA sends in "FWD: My subject" and CC's UserB in that email. UserB replies to that and the subject (depending on email clients) will be "RE: FWD: My Subject" or sometimes just "RE: My Subject", but since the original email subject has the "FWD" prefix, it would not normally match on the subject field.

The new original_subject field stores the subject with all common prefixes stripped off so this sort of scenario is properly handled. We can just scan the original_subject field to find a match.