Don't use UID for dupe checking in ticket gateway.
UIDs are supposed to be guaranteed unique across specific POP3 servers, and this code was properly only checking the specific account. But there is potentially cases where POP3 servers are moved transparently (e.g., new server but exact same credentials so no new gateway is created in DeskPRO), or the server was reset, the server doesn't implement the spec fully, etc. In these sorts of cases we could end up with UID's being the same, and then false positives.

So better to just not use UID here and rely on the usual dupe detection using hashes.