Fix memory usage with custom fields
Main problem here is sharing a $field_group when the form itself is not a real form used on a single form. That is, when a field is created for the sole purpose of rendering the field in multiple places. We'd end up with each field being added to a shared form, then the formView being created over and over again, and each time creating new views with more and more fields.