Rework some timeouts a bit
set_time_limit on Linux does not include non-PHP time (e.g., sockets to send mail, db) but it does count that time on Windows. So it's an unreliable way to handle "bad" cases on Windows because the timeout is breached easily if, for example, the network is slow and sending an email is taking 10 seconds each.

These changes increase the timeout for worker jobs and instead tries to handle jobs that take too long in PHP code. If the process is still going for a long time, the job loop is cancelled (so it can restart fresh the next run). So for quick processes like chat pings can still run regularly even if a job like email collection is taking a long time. The increased set_time_limit is still there as a fallback for very bad cases.