Better handling of bots/cookie-less users, and add a user tokens.
User tokens are to try and catch users that move between pages before the initial DeskPRO request returns. So user goes from page1 to page2 to page3 quickly before the DeskPRO tracking script can be loaded and return with a result. But three separate requests were actually sent to DeskPRO, so three separate visitors were created.

The user token sets a cookie with a random token on the user side as soon as the JS loads, and then the DeskPRO tracking uses the token to try and "find" the user on subsequent pages if no visitor record is specifically set.

In other words, a visitor ID is created on the server-side, a user token is created on the client-side. Both are used to try and tie user pages together.