Move preSend call to just before invoking delegated transports.
DelegatingTransport may modify From based on matched gateway in ticket contexts if a gateway is configured with an alias. preSend automatically sets a return-path based on the From, so we need any changes to From to happen before preSend is called or else we'll end up in a situation where From is the alias and Return-Path is the original address.