Modify way TimeAgo auto-updates time tags.
Before, we'd keep an array of elements that were watched. This is generally good for static pages (e.g., web pages) but for DeskPRO it meant we'd need to properly clean up the array when tags were removed. If we missed doing this, we'd get a collection of time tags that were detached from the dom but in the watch array. This array could get big over time.

Better for us is to just re-query the document every minute and fetch the tags that should be updated. Doing a jQuery query against the document to fetch the tags is easier and less error prone than keeping a watch array up to date and correct.