Fix returning incorrect user records when searching LDAP using a generic input.
Add findIdentityBy* (similar to that found in the AD usersource) to search specific mail/uid fields on input, rather than trying to transform input into a canonical name.

Zend_Ldap has getCanonicalAccountName() that does not actually do any directory searching with the input. It simply looks at the current connection options and converts an input name into an account name. For example, if we have input sarah@foo.com but the directory does not apply to specific domain (like is the case with AD), then the canonical account is created by dropping the domain portion to leave the "username" of sarah. Then DeskPRO would search the directory for the user sarah, find a match, and return that. But of course "sarah" might not be "sarah@foo.com".

This only affects email processing where we do not bind (authenticate) the user.

[FIX] Bug in LDAP email address matching could result in an incoming email from an unknown user being attached to an existing user of the same username that exists in the LDAP directory. Affects helpdesks that use the default DeskPRO usersource in addition to an LDAP usersource.