Improve validation handling
Add custom ClassMetadataFactory that overrides how symfony loads constraints. The problem is that there is no easy way to generate a list of constraint code names. Symfony provides English error messages, but no easy way of getting the same list as a list of error codes instead (suitable for apply error markers to a form).

So we are hijacking the English error message to add a special [code] prefix. Then when it comes time to render the error messages to an API response, we can read those codes from the string.

Also did some work around making it easier to generate validation error reponses, and auto-applying those to angular forms. Just makes handling the whole thing a bit easier.