EmailAccount and EmailAccountManager
New EmailAccount defines email accounts. Accounts can be for incoming, outgoing, or both. This basically replaces the three separate EmailGateway, EmailTransport and EmailGatewayAddress entities. Having these three entities being separate was the cause of a number of bugs and complexities. For example, phantom EmailTransports being left behind, or the complexity in deciding which account to use for a ticket etc.

The EmailAccountManager service is now the central way to work with email accounts. It handles loading them, creating transports, creating the fetchers, finding which account to use for an outgoing email, etc.