Add dpTpl directive so we can significantly improve render time on hot-spots.
dpTpl uses underscore templates to render HTML which work on strings using native JS,
rather than DOM manipulation. This allows very specific control over the template that is output.

The result of the executed template is then in turn compiled by angular so we can still use all
of the nice angular features where required.

Using dpTpl is only necessary in extreme cases. For example, in the ticket list where we would
otherwise be binding thousands of variables and thousands of DOM manipulations. This is slow,
and not even needed because the data does not change very often.