Add other ways to specify test values in trigger term helper methods.
AbstractTriggerTerm helper methods take a prop_name that is used to fetch
the value to be used in a term test. For example, 'subject' means to get
$ticket->subject.

Now, you can use dot-notation to specify values on sub-values. For example,
'person.name' means $ticket->person->name.

Alternatively, you can also use a TermValue class to specify a specific
value that will not be fetched dynamically.