Allow getValueOpArray to return collections, have isStringMatch work on arrays.
The built-in value helpers specify a field to resolve a value to.
For example, person.name is resolved to $ticket->person->name and
the string is returned and used for testing.

Now, you can use square brackets to indicate you want to grab
values from a collection. For example, person.emails[].email
means to return an array of emails.

The goal of using this kind of syntax is to make the default
helpers as expressive as possible so you dont need custom code
to perform common operations.