Allow tracking of non-state related changes to the state recorder.
- Some systems might use the 'version' ID on the state recorder, so if
we know a change is not related to property changes, we can avoid incrementing
the state.

- This is used to track email logs in tickets. So by avoiding incrementing the
version ID when adding a new email log, caches like the ticket filter notify list,
arent reset at the end of the trigger run. Without this, any email log would inc
the version which would make the system think the ticket was chagned by some other
trigger and that we would need to re-run the filter list.