Hack ezc mail parser to keep track of raw source in each part.
ezc parses messages parts into memory as it goes. This is generally good,
but sometimes we want the raw source for an entire part (e.g., an attached rfc822 message).

Before, we re-created a body from the parsed message but this could create imperfect clones
of the original message. Now we are able to get the perfect raw source of the part instead.