[Auth] create agent usersources and move agents to them
this is an upgrade commit, aimed at duplicating the usersources that were setup in the last upgrade commit, but for agents.

Since every person in the system is associated to a user usersource, it will also find all agents that are associated to a user usersource and update that assocation to make it match the agent usersource.