[FIX] "Can use tickets" permission could be deleted in some cases when you wouldn't want it to.
These old cleanup routines incorrectly assume that "can use tickets" shouldn't exist if there
are no departments granted on the same usergroup. So this was assuming that if you didn't
add a usergroup to ANY departments, that's the same as "can use tickets" being off.

This isn't true though because permissions are additive. Departments may be granted through
some other usergroup, but the "can use tickets" itself might just be granted through this one.