Throttle can be a bit higher.
There is a small delay in normal client message polls (which contains the triggers
that make the UI update in the first place).

So the throttle should be large enough to collect a few batches, rather than cause
a situation like "poll, recount, poll, update". We ideally want "poll, poll, poll, recount batch".