Added unit tests for email cutting
Added some basic unit tests for the actual email cutting.

I’ve had to make the cutter language-independent which could cause
issues if a smart user was to insert a string that follows the cutter’s
format.

I’m not really satisfied with that, but all the ways to get the
dependency injection container working in unit tests I’ve seen so far
seem really non-straightforward.