First leg of improvements to Ampliflux:
1. Store hierarchy

Each module now contains it's state under the name of the module.
For example, Modules/Tasks/Reducers/projectList.js and taskList.js
are stored in the stated under Tasks (Tasks.projectList and Tasks.taskList).

TODO: Make it possible to use unlimited depth (i.e., Tasks.x.y.z) and
strategy for connecting components to the state.

2. Refactoring of createAction

* createAction() no longer wraps your function in it's own. If you want
dispatch/getState, then you can wrap you own function to get it.

* Added support for metaFn param which lets you set the 'meta'
value of the action object (a standard FSA property).

* Added support for actionFn being a hard-coded value, so that value
is always returned.

3. Refactoring to reducer helpers

Removed the Reducer class in favour of a createReducer helper function.
The previous method of a reducer class having to be 'compiled' was a bit
hinky.

Now, you pass a build function to createReducer which helps you build
your reducer with helper methods.

4. New middleware

* guid: Every action is assigned a unique ID in action.meta.guid.
Currently unused but perhaps useful in future for logging.

* logger: A logger that console logs event dispatches. Currently
turned on by default, but we might make this some kind of dev option.

* interval: Any action that includes meta.interval will schedule
an interval where the action is dispatched every N ms.

* timeout: Any action that includes meta.delay will schedule
a timeout where the action is dispatched after N ms.

* thunk: Improved thunk middleware handles functions that exist
on action payloads as well. An action payload as a function
is re-dispatched with a special re-dispatch action code. It's
currently only used in the promise middleware to support an
async action that needs to use dispatch.

* promise: Much better promise middleware that now dispatches
START, DONE and ERROR actions (in addition to the usual one).
Combined with createReducer's builder, makes it very easy to
handle loading and error states in response to async actions.