Refactoring and UrlCorrector listener:
- Refactoring to BrandContainer and stuff around themes. BrandContainer
is not a portal-only concept and it will be used in different contexts,
so it makes sense to decouple the portal-specific themeing from it.

- This allowed sharing more services with old DpKernel interfaces. So
PortalKernel and DpKernel are sharing more, and they both can use AppBundle.

- Added UrlCorrector as a symfony kernel event listener (instead of
hard-coded in the kernel). The goal is that the Kernel class can be
much smaller.