Change custom field value column to bigint.
This changes it for new installs, but does skips the change in upgrade.
The change would be very intensive and is not strictly necessary yet
because it only matters for Money fields and only matters if we want
to expose numeric DB functions (e.g. filtering by range) on very large
amounts. So lets defer the col change for future or on an as-needed.