Cloud never has fs blobs.
So if we go through the fs path, we can immediately exit and go through the db lookup path instead. This in turn will redirect to s3 like we want it to do.

This happens if we move a customer from on-prem to cloud. They will have attachments in the fs, which means authcodes with the fs pattern. It's slow to actually invoke the 'move blobs' command, so we just upload their files to s3 and change the file_url and path appropriately. This extra check on the fs code path just optimises away fs existence checks that will never pass on cloud anyway.