Reverts fix for DPLEG-552 that is a big change in behaviour.
https://github.com/deskpro/deskpro/pull/5782

This issue is about letting users create tickets even if they
dont have perm to any deps.

Prior to this PR, this would manifest itself as a form
without a department field at all. The user still has
perm to create tickets, but no deps = no department field.

Tickets always require a department, so in the absence of one,
it will always default to whatever department is set as the system
default.

AFTER this PR, this scenario just breaks with a permission error.
This has had a really big impact on lots of customers so I am
reverting it.

IMHO the old behaviour was correct anyway;
department perms only dictate what the user can select themselves,
it has nothing to do with validating the department on a ticket
after that.