Add source_ref to blobs.
We have imperfect handling of is_temp blobs and the cleanup of is_temp blobs
was commented-out to fix bugs where people's attachments were randomly disappearing.

I've added a new source_ref field to blobs for two reasons:

a) We currently don't have any map of a blob back to its owning object. It makes
sense to try and save at least some sort of id when we know it, at least for auditing.

b) We can audit each place in code where blobs are created, and one by one, we can
confirm if is_temp is being handled properly. If it is, then we ammend the code
to set a source_ref.

c) This means the cleanup script can look for is_temp *and* non-null source_ref.
If both are set, it indicates the blob was uploaded in a place we've confirmed
is handling is_temp properly.

d) Every place where is_temp is NOT validated remains in-place for us to look
at on a case-by-case.

e) The ServeFileScript has been chagned here to stop serving is_temp blobs
after a few hours. If a customer complains of missing images, we have
a fallback where we can just toggle is_temp=0. They won't actually be deleted,
so no data loss.